Portable library models. * E.g., for <string.h> and <math.h> * Avoids conflicts with system headers by hijacking include path. * Replacement header for `$MODULE` in `share/smack/include/$MODULE/$MODULE.h`. * Replacement model for `$MODULE` in `share/smack/lib/$MODULE.c`. * Only use custom headers when models are used too. * Fixes #238.
